+++ src/common/include/common.h
@@ -12,7 +12,7 @@ bool InvalidPort(const std::string& port);
 int StringAsInt(const std::string& str);
 std::string GetPeerInfo(struct sockaddr_in *);
 std::string GetPeerIP(struct sockaddr_in *paddr);
-int GetPeerPort(struct sockaddr_in *paddr);
+unsigned short GetPeerPort(struct sockaddr_in *paddr);
 std::string i2s(const int& i);
 int s2i(const std::string& s);
